C99 testsuite readiness: Verified un-reductions
	* gcc.c-torture/compile/20080613-1.c (hop_sendmsg): Call
	__builtin_memcpy instead of memcpy.
	* gcc.c-torture/compile/complex-6.c (bar): Declare.
	* gcc.c-torture/compile/pr23445.c (__brelse): Declare.
	* gcc.c-torture/compile/pr23946.c (long2str): Declare.
	(avi_parse_comments): Call __builtin_memset instead of
	memset.  __builtin_malloc instead of malloc.  Call
	__builtin_memcpy instead of memcpy.  Call
	__builtin_free instead of free.
	* gcc.c-torture/compile/pr31953.c (toggle_hexedit_mode):
	Add missing void return type.
	* gcc.c-torture/compile/pr32372.c (MPV_encode_init): Add
	missing void return type.
	* gcc.c-torture/compile/pr32355.c (sigemptyset): Declare.
	(doSignalsSetup): Add missing void return type.
	* gcc.c-torture/compile/pr32453.c (__assert_fail): Declare.
	* gcc.c-torture/compile/pr32571.c (mthca_is_memfree)
	(mthca_arbel_fmr_unmap, mthca_tavor_fmr_unmap)
	(mthca_unmap_fmr): Declare.
	* gcc.c-torture/compile/pr32584.c (sortpin): Add missing
	void types.
	* gcc.c-torture/compile/pr32919.c (read_int, _itoa_word)
	(__strnlen): Declare.
	* gcc.c-torture/compile/pr33173.c (test_dir_format): Add
	missing void return type.  Add missing int types.
	* gcc.c-torture/compile/pr33855.c (cabsl): Declare.
	* gcc.c-torture/compile/pr34334.c (__strsep_1c)
	(__strsep_2c): Add missing void return type.
	* gcc.c-torture/compile/pr35006.c (grub_putchar)
	(cl_set_pos, cl_print, grub_memmove, cl_delete): Declare.
	(grub_cmdline_get): Add missing void return type.
	* gcc.c-torture/compile/pr35595.c (__kernel_sinf):
	Declare.
	* gcc.c-torture/compile/pr35869.c (f): Add missing void
	return type.
	* gcc.c-torture/compile/pr36172.c (FcCharSetFreeze): Add
	missing return value.
	* gcc.c-torture/compile/pr36238.c (lshift_s_s): Declare.
	* gcc.c-torture/compile/pr37207.c (func_81, func_98):
	Declare.
	* gcc.c-torture/compile/pr37258.c (mod_rhs, lshift_s_s)
	(func_112, func_23): Declare.
	* gcc.c-torture/compile/pr37305.c (safe_mod_u_u): Declare.
	* gcc.c-torture/compile/pr37327.c (func_93, func_59)
	(func_124, func_117, safe_add_uint64_t_u_u)
	(safe_mul_int32_t_s_s): Declare.
	* gcc.c-torture/compile/pr37387.c (FuncMakeConsequencesPres):
	Call __builtin_abort instead of ErrorQuit.
	* gcc.c-torture/compile/pr37432.c (print_wkb_bytes): Declare.
	* gcc.c-torture/compile/pr37713.c (sdp_seq_alloc): Declare.
	* gcc.c-torture/compile/pr39886.c (func): Declare.
	* gcc.c-torture/compile/pr39941.c (stop): Declare.
	* gcc.c-torture/compile/pr41016.c (CompareRNAStructures):
	Call __builtin_abort instead of Die.
	* gcc.c-torture/compile/pr42632.c (___pskb_trim): Add
	forward declaration.
	* gcc.c-torture/compile/pr49710.c (baz): Add forward
	declaration and missing void types.
	(bar): Add missing void type.
	* gcc.c-torture/compile/pr52437.c (fn2): Declare.
	* gcc.c-torture/compile/pr57441.c (func_1): Add missing void
	return type.
	* gcc.c-torture/compile/pr87110.c (struct d): Add missing
	semicolon.
	(g, h): Define as int.
	(i): Add missing void types.
	* gcc.c-torture/compile/pr87468.c (a): Define as int.
	(e, f): Declare.
	(b): Add missing void types.
	* gcc.c-torture/execute/pr79043.c (ptr2): Use cast in
	initializer.
	(typepun): Add missing void return type.
	(main): Add missing int return type and missing void type.
	* gcc.dg/pr100349.c (b): Add missing void return type.
	* gcc.dg/pr106189.c (T): Declare.
	* gcc.dg/pr110777.c (_setjmp): Declare
	* gcc.dg/pr45506.c (dynvec, relative_relocp, atexit): Declare.
	(fini_array): Cast result of relative_relocp from int to int *.
	* gcc.dg/pr97359.c: Compile with -Wno-pointer-to-int-cast.
	(b): Cast point to int to avoid int-conversion warning.
	* gcc.dg/uninit-pr78548.c: Call __builtin_printf instead
	of printf.
	* gcc.dg/torture/pr39829.c (f): Declare.
	* gcc.dg/torture/pr44977.c (int329): Cast bar from pointer
	to short.
	* gcc.dg/torture/pr53703.c (ifa_sa_len): Declare.
	(usagi_getifaddrs): Call __builtin_memset instead of memset
	and __builtin_memcmp instead of memcmp.
	* gcc.dg/torture/pr68625.c (z9): Explicitly cast
	pointers to int.
	* gcc.dg/torture/pr55964.c (f): Cast q to the expected type.
	* gcc.dg/torture/pr70623.c (h9): Fix pointer cast in assignment
	of av.
	* gcc.dg/torture/pr81118.c (c): Return zero instead of
	nothing.
	* gcc.dg/torture/pr81510.c (e): Add cast from int to pointer.
	* gcc.dg/torture/pr99694.c (d): Add missing casts between
	pointers and integers.
	* gcc.dg/tree-ssa/pr102563.c: Define as void (*) (void).
	(_bdf_parse_glyphs_p): Define with prototype.
	* gcc.dg/tree-ssa/pr92085-1.c (di): Add cast to initialization
	of fq.
	* gcc.dg/tree-ssa/pr92085-2.c (c1): Add casts to
	initializations of th, lq.
	* gcc.dg/tree-ssa/reassoc-43.c
	(c_parser_next_token_is_keyword, chainon)
	(c_parser_attributes): Declare.
	* gcc.dg/tree-ssa/split-path-4.c (powi_lookup_cost): Declare.
	(powi_cost): Add missing void return type.
	* gcc.dg/tree-ssa/split-path-6.c (strtosichar)
	(ichartosstr, treeinsert, stop, __srget_r): Declare
	(givehelp, oof, lookharder): Convert to prototypes.
	* gcc.target/i386/pr71621-1.c (y8): Cast to expected
	pointer type in assignment of t6.
	* gcc.target/i386/pr79571.c (g): Add missing int type.
	(j): Add missing void types and pointer cast.
	* gcc.target/i386/pr87246.c (s2): Add cast from integer
	to pointer type.
	* gcc.target/i386/pr87759.c (rc): Cast to the expected
	pointer time in initialization of ar.
	* gcc.target/i386/pr99454.c (tg3_start_xmit): Add
	missing cast from int to int *.

+++ gcc/testsuite/gcc.dg/torture/pr68625.c
@@ -7,7 +7,7 @@ int sg;
 void
 z9(void)
 {
-  int pz, oi, vz, yp, zi, hd, pw, gr, w9 = 0, j0 = -1, rb = &w9;
+  int pz, oi, vz, yp, zi, hd, pw, gr, w9 = 0, j0 = -1, rb = (int) &w9;
   int *lr;
   while (w9 < 1) {
       lr++;
@@ -24,7 +24,7 @@ ee:
       **dp = 1;
       while (**dp)
 	if (++oi) {
-	    int mq = dp;
+	    int mq = (int) dp;
 	    j0 = 1;
 	    while (pw < 1) {
 		if (++rb && mq)
